requirement.txt
faiss-cpu
groq
langchain_groq
pyPDF2
langchain_google_genai
langchain
streamlit
langchain_community
python-dotenv
pyPDF